requirements.txt
click==8.1.3
matplotlib==3.7.5
mlflow==2.15.1
mlflow_skinny==2.15.1
azureml-mlflow
numpy==1.24.4
pandas==2.0.3
scikit_learn==1.3.2
seaborn==0.13.2
statsmodels==0.14.1
zenml==0.64.0
xgboost
streamlit
-e .